largest-contentful-paint/first-letter-background.html

<!doctype html>
<meta charset="utf-8" />
<title>Largest Contentful Paint: observe element with background image in its first letter</title>
<style>
  div::first-letter {
    background-image: url("/images/black-rectangle.png");
  }
  div {
    font-size: 12px;
  }
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/largest-contentful-paint-helpers.js"></script>
<body></body>
<script>
  setup({ hide_test_state: true });
  promise_test(async (t) => {
    assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");
    const entries = [];
    let observer = new PerformanceObserver((list) => {
      entries.push(...list.getEntries());
    });
    observer.observe({ entryTypes: ["largest-contentful-paint"] });
    const beforeLoad1 = performance.now();
    document.body.innerHTML = "<div id='target'>A</div>";
    await t.step_wait(
      () => entries.some((entry) => entry.url.endsWith("black-rectangle.png")),
      "First-letter background image loaded.",
    );
    const beforeLoad2 = performance.now();
    const div = document.createElement("div");
    div.innerText = "The long text will now be LCP.";
    div.id = "target2";
    document.body.appendChild(div);
    await t.step_wait(() => entries.some((entry) => entry.id === "target2"), "long text is LCP.");
    observer.disconnect();
    if (entries.length === 3) {
      // Check the short text (non background). This one may not appear.
      const entry = entries.shift();
      assert_equals(entry.url, "", "expected text to be LCP");
      assert_equals(entry.id, "target");
    }
    // At this point we should have 2 entries: the first-letter background image and the long text.
    assert_equals(entries.length, 2, "expected 2 entries");
    {
      // Check the first-letter background image of the short text.
      const entry = entries.shift();
      assert_equals(entry.entryType, "largest-contentful-paint");
      const url = window.location.origin + "/images/black-rectangle.png";
      checkImage(entry, url, "target", 0, beforeLoad1, ["sizeLowerBound"]);
    }
    {
      // Check the long text.
      const entry = entries.shift();
      assert_equals(entry.entryType, "largest-contentful-paint");
      assert_greater_than_equal(entry.renderTime, beforeLoad2, "blaaa");
      assert_greater_than_equal(performance.now(), entry.renderTime, "bleee");
      assert_approx_equals(
        entry.startTime,
        entry.renderTime,
        0.001,
        "startTime should be equal to renderTime to the precision of 1 millisecond.",
      );
      assert_equals(entry.duration, 0);
      assert_equals(entry.id, "target2");
      const div = document.getElementById("target2");
      // Estimate the text size: 12 * 100
      assert_greater_than_equal(entry.size, 1200);
      assert_equals(entry.loadTime, 0);
      assert_equals(entry.element, div);
    }
  }, "Largest Contentful Paint: first-letter is observable.");
</script>
